+++ parachain/modules/consensus/polygon-pos/prover/src/test.rs
@@ -1,6 +1,7 @@
 use ethers::providers::{Provider, Ws};
+use geth_primitives::Header;
 use ismp::util::Keccak256;
-use polygon_pos_verifier::{primitives::Header, verify_polygon_header};
+use polygon_pos_verifier::verify_polygon_header;
 
 use crate::PolygonPosProver;
 
@@ -30,12 +31,12 @@ async fn verify_polygon_pos_headers() {
 
     let (mut finalized_header, mut validators) = prover.fetch_finalized_state().await.unwrap();
 
-    let mut parent_hash = Header::from(&finalized_header).hash::<Host>().unwrap();
+    let mut parent_hash = Header::from(&finalized_header).hash::<Host>();
     for number in (finalized_header.number.low_u64() + 1)..=(finalized_header.number.low_u64() + 10)
     {
         let header = prover.fetch_header(number).await.unwrap().unwrap();
         if parent_hash == header.parent_hash {
-            parent_hash = Header::from(&header).hash::<Host>().unwrap();
+            parent_hash = Header::from(&header).hash::<Host>();
             let result = verify_polygon_header::<Host>(&validators, header).unwrap();
             finalized_header = result.header;
             if let Some(next_validators) = result.next_validators {
