+++ parachain/modules/consensus/polygon-pos/verifier/src/lib.rs
@@ -2,8 +2,9 @@
 #[warn(unused_imports)]
 #[warn(unused_variables)]
 use anyhow::anyhow;
+use geth_primitives::{CodecHeader, Header};
 use ismp::util::Keccak256;
-use primitives::{get_signature, parse_validators, CodecHeader, Header};
+use primitives::{get_signature, hash_without_sig, parse_validators};
 use sp_core::{H160, H256};
 pub mod primitives;
 use alloc::collections::BTreeSet;
@@ -23,13 +24,13 @@ pub fn verify_polygon_header<I: Keccak256>(
     let signature = get_signature(&header.extra_data)?;
     let next_validators = parse_validators(&header.extra_data)?;
     let rlp_header: Header = (&header).into();
-    let msg = rlp_header.clone().hash_without_sig::<I>()?;
+    let msg = hash_without_sig::<I>(rlp_header.clone())?;
     let address = sp_io::crypto::secp256k1_ecdsa_recover(&signature, &msg.0)
         .map_err(|_| anyhow!("Signature verification failed"))?;
     let signer_hash = H160::from_slice(&I::keccak256(&address[..]).0[12..]);
     if !validators.contains(&signer_hash) {
         Err(anyhow!("Header is signed by unknown validator"))?
     }
-    let hash = rlp_header.hash::<I>()?;
+    let hash = rlp_header.hash::<I>();
     Ok(VerificationResult { hash, header, next_validators })
 }
